Add イソリア after name so that Apple Maps will pick it up...
Want people to be able to search apple maps for name, but, you can do only one
entry. If you put it as bilingual like "eSolia イソリア", then they return an
error unless that string is locate-able. So, add it to one of the paras to
see if it will work.

Signed-off-by: Rick Cogley <rick.cogley@esolia.co.jp>