Change uptime protocol from secure to non-secure (lol)
So we changed to the excellent "ping" monitor service from @apex (https://apex.sh) but, if you try to use a cname to serve the status pages on your own domain, you get the "ermagah this site is trying to steal your data" error in modern browsers, because the cert for the status pages doesn't cover your domain. So if you can accept the "red lock of shame" indicator for unsecured sites these days, you can share status pages on your own domain, as httP urls.

Signed-off-by: Rick Cogley <rick.cogley@esolia.co.jp>